Move unauthorized template to its own file
